# Tweak a few things to improve performance